request refactor

+++ script.js
@@ -100,11 +100,11 @@ pushNotification({
   orderId: "order-1",
   token:
     "fml04QoeSGqzdJ5CwyXxES:APA91bE9iP6bSSsasYBtU4Z6BBPbrS6JKi6J0-LoLebypSuKYPgKVnAOCII8V2oqIlWlRQemMEKil7ja4zswjVwprtQMWzto12ayml5qLOolDoBgicMnemzJWdEt2vpEgowQMY38oLq1",
-  data: { Value: parseInt(total) },
+  data: { Value: parseFloat(total) },
   notification: {
     title: `${comerciante}`,
     message: `Pagamento de €${total}`,
   },
   telemovel: `${phoneNumber.value}`,
-  // details (optional): [{name: 'Product 1', size: 'M', quantity: '1', price: '10'}]
+  // details (optional): [{name: 'Product 1', size: 'M', quantity: '1', price: '10'}, ...]
 });
